server/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
apitally==0.11.0
backoff==2.2.1
bcrypt==4.0.1
bidict==0.23.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.5.0
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
ez_setup==0.9
fastapi==0.110.0
filelock==3.13.1
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
kafka-python==2.0.2
kafka-python-ng==2.2.2
Mako==1.3.2
MarkupSafe==2.1.5
msgpack==1.0.8
passlib==1.7.4
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==57.5.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sse-starlette==2.1.2
starlette==0.36.3
typing==3.7.4.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
uvicorn==0.27.1
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
apitally==0.11.0
backoff==2.2.1
bcrypt==4.0.1
bidict==0.23.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.5.0
coverage==7.6.0
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
ez_setup==0.9
fastapi==0.110.0
filelock==3.13.1
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
kafka-python==2.0.2
kafka-python-ng==2.2.2
Mako==1.3.2
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.1
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==57.5.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sse-starlette==2.1.2
starlette==0.36.3
typing==3.7.4.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
uvicorn==0.27.1
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
apitally==0.11.0
backoff==2.2.1
bcrypt==4.0.1
bidict==0.23.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.5.0
coverage==7.6.0
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
ez_setup==0.9
fastapi==0.110.0
filelock==3.13.1
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
kafka-python==2.0.2
kafka-python-ng==2.2.2
Mako==1.3.2
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.1
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==57.5.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sse-starlette==2.1.2
starlette==0.36.3
typing==3.7.4.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
uvicorn==0.27.1
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
